fix appdata parsing
I don't think the statement that open-uri decompresses automatically
is correct (and it seems not to work). The code does not reference
anything like that and probably the code is confusing Content-Type
with Content-Encoding. So, fix it to handle decompression correctly.

This makes fill_search_cache work again and screenshots appear in my
local copy that are not in software.opensuse.org current deploy.