Replace PhantomJS with Headless Firefox
When Headless Chrome was released, PhantomJS was deprecated/abandoned, as it did not make sense anymore.

Since Firefox 56, headless mode is also supported, and there is a nifty gem that installs the binary of the selenium webdriver for you.

Depends on Rails 5.2, due to: https://github.com/rails/rails/issues/29688
(fix not working here on Rails 5.1.6)

Also, the tests that need access to the status code stay as integration tests, as
Selenium Webdriver does not support access to the status code:
https://github.com/SeleniumHQ/selenium-google-code-issue-archive/issues/404