Do not allow screenshot source urls as input parameters
The end-point /package/screenshot/fakename.png?appscreen=$url allows any caller to make the app
download files of any size.

Reported by Espartaco Palma <esparta@gmail.com>

In order to remove this flaw, the following refactoring is done:

* Move all code that gets the thumbnails to a single end-point

The /package/screenshot /package/thumbnail will redirect to the original image, and to
the generated thumbnail. The end-point now only takes the package name, so the url comes from the appdata, and not from an input parameter.

* Remove the memcached layer for screenshots

The screenshots were cached both on disk and in memcached, without a clear reason. Adding a memcached layer is now possible in a more predictable fashion, in case it is needed.

* Do not read the content into memory.

Except for the initial download performed by MiniMagick contructor, there is no reading of image data into memory anymore. Only the urls are generated, and the end-point redirects to both the default images (served from the asset pipeline), and the generated thumbnails (served from public/).