Update docker environment to Leap 15 + ruby2.5
Use latest leap for docker and update documentation.
Installing `rbtrace` in the docker environment does not succeed and is
disabled in the Gemfile.

Docker CMD line is removed since gems are not installed automatically,
so the first `vagrant up` does not start rails and no further
provisioning is possible without SSH.