Add OBS mini client
Use a mini client for connections to our backend OBS. This client is
based on Faraday, a flexible HTTP client library. It is easily
extensible by Rack middleware and hopefully makes it easier to debug
memory issues. It will replace Seeker, the currently used model that
handles these queries, together with ActiveXML.

Co-authored-by: Duncan Mac-Vicar P <dmacvicar@suse.de>