Revamp download section
Changes to the way info regarding distributions is generated, instead of
creating 3 seperate templates for distributions, now it's one file with
seperate data files feeding into it (in app/data directory).

Distributions now also are properly colour coded in every step of the
way, from /distributions to specific distribution's page.

Kubic was also moved into seperate tab in Tumbleweed distro page.

Speaking of tabs, they are now more visible against $distrocolor
background, so it's harder to miss.

Moreover, buttons in those tabs are now way more visible, so user knows
which button is for primary download and which buttons function as
alternatives of the same download.

Lastly, elements of styling and images that went unused since bento were
removed to unclutter the repository.