Run Travis for all branches except depfu
According to de04c9dcb51315cc2ae42cfe771e2d0b8be207ee, Travis is only
needed in the master branch, but I would also like to run it for my
fork. This can be enabled in the user options in Travis, but it is
rejected because of the branch being excluded per configuration.

depfu branches keep excluded as it will otherwise cause that we run it
twice.