Handle connection error in retrieve_image_size
FaradayMiddleware::FollowRedirects has a default limit of 3 redirects.
This is not always enough with download.opensuse.org's mirrorbrain
infrastructure. A limit of 5 should work around those cases.

If there are still too many redirects or any other HTTP connection
error, a default of "0" is set instead of raising the error which would
stop rendering the page.