Move image_size logic from view to helper
There is no need to save the image_size in the view, just passig the
medium to the helper is enough. `short_description` is called with just
a medium (or "type") as the argument, the method finds out on its own if
it needs to image_size or not.

The view reads a bit clearer this way.