Change how baseproject is set up
The baseproject can be set either via the Cookie header or via the query
string. The latter takes precedence over the cookie value, i.e. opening
a search URL that contains a baseproject key will override the
previously set cookie. This helps with the sharing of searches.

The default baseproject is the latest Leap stable release.