Use constraints for /download routes
Actions don't need to validate if all required parameters are provided
if they are not reached in the first place. Route validations are not a
concern of the action, instead they should be done in config/routes.rb.
This also allows us to remove required_parameters, which causes a lot of
false positive "Unpermitted parameters" log statements.

As a result of using route constraints, no error flashes of this form

   "Required Parameter foo missing"

are shown anymore. While we show /download/doc instead, this page is not
specifically pointing to the problem. I don't see this as a big problem
since the flash was lacking already context, users had to guess what a
"Parameter" is and how it should be provided. /download/doc does not
worsen the situation substantially.