Remove required_parameters function
This function provided a way to ensure that the params hash contains
specific keys. The function is not needed if routes are validated
properly using constraints. As a side effect, required_parameters caused
a lot of false positive "unpermitted parameters" log statements.

required_parameters is not used anywhere in the codebase anymore.

Fixes https://github.com/openSUSE/software-o-o/issues/500.