Fix tabs in distribution page
Do no use translations as tabs. Use `tab-<number>` instead.

Another solution is to have the English names in the version files and
use the translations when needed. This makes a bit easier to modify
the version files and would allow us to have nicer links. But the change
is a bit more complicated and this page may be migrated to:
https://get.opensuse.org

Fix (partially): https://github.com/openSUSE/software-o-o/issues/596