package.erb: Update Arch Linux instructions to support signature verification
OBS has supported both signed metadata and packages for Arch Linux for a
while now. This commit updates the setup instructions to enable
signature verfication.

The steps for Arch Linux are a bit more complicated than other distros.
The GPG public key must first be imported into the pacman GPG keyring
and then be locally signed to trust the key.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>