Don't search with baseproject for Leap 15.3
Having a baseproject in the query does not work well for Leap 15.3 due
to the way it is built. There are multiple "baseprojects", which are now
specified in the distributions_projects.rb initializer.

Once the query to OBS returns, the correct baseproject
(openSUSE:Leap:15.3) is added to packages that should have it.