Improve the Debian/Ubuntu instructions for adding the OBS repository
I has come to my attention that the `apt-key` practice
(as well as the `trusted.gpg.d`) are not recommended by the Debian team
(https://wiki.debian.org/DebianRepository/UseThirdParty#OpenPGP_Key_distribution),
and instead `deb [ signed-by=... ]` is the recommended practice.

I cannot find an official source on why is that in "nice terms", but this article
https://michael-prokop.at/blog/2021/02/16/how-to-properly-use-3rd-party-debian-repository-signing-keys-with-apt/
gives more insight into the issue.

Additionally, for Ubuntu `wget` > `curl`, since `curl` needs to be installed separately.

The issue described in the `h6` above is still "relatively true", but it's less severe.

Tested in a home-brewed test, based on the
https://software.opensuse.org/download.html?project=home:vtrefny&package=blivet-gui
instructions, inside an Ubuntu docker container (with the necessary setup).

Signed-off-by: Stavros Ntentos <133706+stdedos@users.noreply.github.com>

Since we are not using `-O <filepath>`, `sudo wget` is not needed

Oops 😅

`gpg --dearmor` is a little bit unnecessary on nowadays `SecureApt`.
However, let's not change this to preserve backwards-compatibility instead.

Signed-off-by: Stavros Ntentos <133706+stdedos@users.noreply.github.com>